DMT-64 Add extra control flow and error handling blocks
Add extra error checking capabilities mostly in terms of outputting corresponding error messages to the user, depending on what the error was and where it was caught to main.js and donutState.js.
Add more error messages to the resources.js
Add a check in renderer.js in regard to when the height or width of the canvas container becomes too small.