DMT-49 Update the calculation of the sectors' label text
Updated the logic of the calculation of the sectors' label text, to return the same color as the background color of the canvas (in case of the corresponding contrast ratio is low); this way a more similar to the pie chart approach has been observed.
In addition, comments about the newly introduced function has been added, along with any correlated reference sources used.